[change] move db and models to their own modules
