+++ pyproject.toml
@@ -12,6 +12,7 @@ python = "3.9.16"
 ata-db-models = "0.0.17"
 fastapi = "^0.95.0"
 mangum = "^0.17.0"
+pydantic = "^1.10.8"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^0.961"
